+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Yii Framework 2 swiftmailer extension Change Log
 ================================================
 
-2.0.6 under development
------------------------
+2.0.6 September 09, 2016
+------------------------
 
 - Enh #6: Added ability to specify custom mail header at `yii\swiftmailer\Message` (klimov-paul)
 - Enh #23: Added `yii\swiftmailer\Message::setReturnPath()` shortcut method (klimov-paul)